+++ src/components/BurnerWalletCard/index.tsx
@@ -4,13 +4,15 @@ import { useState } from 'preact/hooks'
 import BurnerWalletDivider from 'icons/BurnerWalletDivider'
 import BurnerWalletHeaderLogo from 'components/BurnerWalletCard/BurnerWalletHeaderLogo'
 import Button from 'components/ui/Button'
-import CardContainer from 'components/BurnerWalletCard/CardContainer'
 import ColorfulEye from 'icons/ColorfulEye'
 import Copy from 'icons/Copy'
 import CreateFirstCastButton from 'components/BurnerWalletCard/CreateFirstCastButton'
+import MagicCardWrapper from 'components/ui/MagicCardWrapper'
 import classnames, {
   alignItems,
+  backgroundColor,
   blur,
+  borderRadius,
   cursor,
   display,
   flexDirection,
@@ -21,6 +23,7 @@ import classnames, {
   userSelect,
   width,
   wordBreak,
+  zIndex,
 } from 'classnames/tailwind'
 
 const container = classnames(
@@ -63,6 +66,17 @@ const privateKeyBlur = (closed?: boolean) =>
     blur({ 'blur-sm': closed }),
     transitionProperty('transition-all')
   )
+const cardContainer = classnames(
+  display('flex'),
+  flexDirection('flex-col'),
+  gap('gap-y-4'),
+  alignItems('items-center'),
+  borderRadius('rounded-2xl'),
+  wordBreak('break-words'),
+  zIndex('z-20'),
+  backgroundColor('bg-primary-background'),
+  padding('py-6', 'px-6', 'xs:py-10')
+)
 
 export default function () {
   const [isKeyClosed, setKeyClosed] = useState(true)
@@ -71,57 +85,59 @@ export default function () {
   )
 
   return (
-    <CardContainer>
-      <BurnerWalletHeaderLogo />
-      <div className={container}>
-        <HeaderText center>Your new wallet:</HeaderText>
-        <div className={privateContainer}>
-          <div className={privateKeyText}>
-            <BodyText center semiBold>
-              Private key
-            </BodyText>
-            <div className={privateKeyBlur(isKeyClosed)}>
-              <AccentText color="text-secondary">
-                <BodyText center semiBold inheritColor>
-                  {privateKey}
-                </BodyText>
-              </AccentText>
+    <MagicCardWrapper>
+      <div className={cardContainer}>
+        <BurnerWalletHeaderLogo />
+        <div className={container}>
+          <HeaderText center>Your new wallet:</HeaderText>
+          <div className={privateContainer}>
+            <div className={privateKeyText}>
+              <BodyText center semiBold>
+                Private key
+              </BodyText>
+              <div className={privateKeyBlur(isKeyClosed)}>
+                <AccentText color="text-secondary">
+                  <BodyText center semiBold inheritColor>
+                    {privateKey}
+                  </BodyText>
+                </AccentText>
+              </div>
+            </div>
+            <div className={buttonsContainer(isKeyClosed)}>
+              <div
+                className={buttonContent}
+                onClick={() => setKeyClosed(!isKeyClosed)}
+              >
+                <ColorfulEye open={isKeyClosed} />
+                <Button small gradientFont type="tertiary">
+                  {isKeyClosed ? 'Reveal key' : 'Hide key'}
+                </Button>
+              </div>
+              <div
+                className={buttonContent}
+                onClick={async () => {
+                  await navigator.clipboard.writeText(privateKey)
+                  toast('Private key copied 🎉')
+                }}
+              >
+                <Copy />
+                <Button small gradientFont type="tertiary">
+                  Copy key
+                </Button>
+              </div>
             </div>
           </div>
-          <div className={buttonsContainer(isKeyClosed)}>
-            <div
-              className={buttonContent}
-              onClick={() => setKeyClosed(!isKeyClosed)}
-            >
-              <ColorfulEye open={isKeyClosed} />
-              <Button small gradientFont type="tertiary">
-                {isKeyClosed ? 'Reveal key' : 'Hide key'}
-              </Button>
-            </div>
-            <div
-              className={buttonContent}
-              onClick={async () => {
-                await navigator.clipboard.writeText(privateKey)
-                toast('Private key copied 🎉')
-              }}
-            >
-              <Copy />
-              <Button small gradientFont type="tertiary">
-                Copy key
-              </Button>
-            </div>
-          </div>
+          <BodyText center primary semiBold>
+            <AccentText color="text-accent">Save this key—seriously</AccentText>
+            . Otherwise, you’ll lose it to the endless digital void where no
+            wallets return from. This wallet is generated locally, so we will
+            never know this key, and it won’t be saved if you leave this
+            page—kinda sad if you think about it.
+          </BodyText>
+          <CreateFirstCastButton />
+          <BurnerWalletDivider />
         </div>
-        <BodyText center primary semiBold>
-          <AccentText color="text-accent">Save this key—seriously</AccentText>.
-          Otherwise, you’ll lose it to the endless digital void where no wallets
-          return from. This wallet is generated locally, so we will never know
-          this key, and it won’t be saved if you leave this page—kinda sad if
-          you think about it.
-        </BodyText>
-        <CreateFirstCastButton />
-        <BurnerWalletDivider />
       </div>
-    </CardContainer>
+    </MagicCardWrapper>
   )
 }
