+++ src/index.css
@@ -109,6 +109,7 @@ circle {
   left: 0;
   width: 100%;
   height: 100%;
+  z-index: -1;
   background-image: linear-gradient(90deg, #ff7bed, #fed823);
   border-radius: 16px;
   background-size: 200% 200%;
@@ -123,11 +124,13 @@ circle {
   right: 0;
   height: 100%;
   width: 100%;
+  z-index: -1;
   transform: scale(0.8);
   filter: blur(calc(65vh / 6));
   background-image: linear-gradient(90deg, #ff7bed, #fed823);
   background-size: 200% 200%;
-  animation: spin-background 3.5s ease-in-out infinite;
+  animation: spin-background 3.5s ease-in-out infinite,
+    rescale 3.5s ease-in-out infinite;
 }
 
 @keyframes spin-background {
@@ -137,5 +140,14 @@ circle {
   }
   50% {
     background-position: 100% 50%;
+  }
+}
+@keyframes rescale {
+  0%,
+  100% {
+    transform: scale(0.8);
+  }
+  50% {
+    transform: scale(0.85);
   }
 }
